requirements.txt
absl-py==0.1.10
bleach==1.5.0
bluelens-log==1.0.2
bluelens-spawning-pool==1.0.10
boto3==1.4.7
botocore==1.7.47
certifi==2018.1.18
docutils==0.14
enum34==1.1.6
futures==3.1.1
grpcio==1.7.0
grpcio-tools==1.8.4
html5lib==0.9999999
jmespath==0.9.3
Markdown==2.6.9
numpy==1.13.3
olefile==0.44
Pillow==4.3.0
protobuf==3.4.0
pymongo==3.6.0
python-dateutil==2.6.1
redis==2.10.6
s3transfer==0.1.11
six==1.11.0
stylelens-feature==1.0.11
stylelens-image==0.0.9
stylelens-object==0.0.40
stylelens-product==1.0.67
stylelens-s3==0.0.2
tensorflow==1.5.0
tensorflow-tensorboard==1.5.0
urllib3==1.22
Werkzeug==0.12.2
